+++ inc/doc.inline.h
@@ -106,12 +106,22 @@ static inline const char * doc_map(ti_val_t * val)
     }
 }
 
-static inline const char * doc_id_list(ti_val_t * val)
+static inline const char * doc_map_id(ti_val_t * val)
 {
     switch ((ti_val_enum) val->tp)
     {
-    case TI_VAL_ARR:            return DOC_LIST_ID_LIST;
-    case TI_VAL_SET:            return DOC_SET_ID_LIST;
+    case TI_VAL_ARR:            return DOC_LIST_MAP_ID;
+    case TI_VAL_SET:            return DOC_SET_MAP_ID;
+    default:                    return NULL;
+    }
+}
+
+static inline const char * doc_map_wrap(ti_val_t * val)
+{
+    switch ((ti_val_enum) val->tp)
+    {
+    case TI_VAL_ARR:            return DOC_LIST_MAP_WRAP;
+    case TI_VAL_SET:            return DOC_SET_MAP_WRAP;
     default:                    return NULL;
     }
 }