+++ inc/ti/fn/fnmapid.h
@@ -1,6 +1,6 @@
 #include <ti/fn/fn.h>
 
-static int id_list__walk_set(ti_thing_t * thing, ti_varr_t * varr)
+static int map_id__walk_set(ti_thing_t * thing, ti_varr_t * varr)
 {
     ti_val_t * val = thing->id
             ? (ti_val_t *) ti_vint_create((int64_t) thing->id)
@@ -12,18 +12,18 @@ static int id_list__walk_set(ti_thing_t * thing, ti_varr_t * varr)
     return 0;
 }
 
-static int do__f_id_list(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+static int do__f_map_id(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 {
     const char * doc;
     const int nargs = fn_get_nargs(nd);
     size_t n;
     ti_varr_t * varr;
 
-    doc = doc_id_list(query->rval);
+    doc = doc_map_id(query->rval);
     if (!doc)
-        return fn_call_try("id_list", query, nd, e);
+        return fn_call_try("map_id", query, nd, e);
 
-    if (fn_nargs("id_list", doc, 0, nargs, e))
+    if (fn_nargs("map_id", doc, 0, nargs, e))
         return e->nr;
 
     n = ti_val_get_len(query->rval);
@@ -40,9 +40,9 @@ static int do__f_id_list(ti_query_t * query, cleri_node_t * nd, ex_t * e)
             if (!ti_val_is_thing(val))
             {
                 ex_set(e, EX_TYPE_ERROR,
-                        "function `id_list` requires a list with items of "
-                        "type `thing` only but got type `%s` instead"
-                        DOC_LIST_ID_LIST,
+                        "function `map_id` requires a list with items of "
+                        "type `thing` but found an item of type `%s` instead"
+                        DOC_LIST_MAP_ID,
                         ti_val_str(val));
                 goto fail0;
             }
@@ -60,7 +60,7 @@ static int do__f_id_list(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         }
         break;
     case TI_VAL_SET:
-        if (imap_walk(VSET(query->rval), (imap_cb) id_list__walk_set, varr))
+        if (imap_walk(VSET(query->rval), (imap_cb) map_id__walk_set, varr))
         {
             ex_set_mem(e);
             goto fail0;
@@ -73,6 +73,6 @@ static int do__f_id_list(ti_query_t * query, cleri_node_t * nd, ex_t * e)
     return e->nr;
 
 fail0:
-    ti_val_drop((ti_val_t *) varr);
+    ti_val_unsafe_drop((ti_val_t *) varr);
     return e->nr;
 }